remove redundant synchronize, add clear command, and update code structure
