+++ lmcache/v1/multiprocess/futures.py
@@ -1,0 +1,57 @@
+# SPDX-License-Identifier: Apache-2.0
+# Standard
+from typing import Generic, Optional, TypeVar
+import threading
+
+T = TypeVar("T")
+
+
+class MessagingFuture(Generic[T]):
+    def __init__(self):
+        self.is_done_ = threading.Event()
+        self.result_ = None
+
+    def query(self) -> bool:
+        """
+        Check if the future is done.
+
+        Returns:
+            bool: True if the future is done, False otherwise.
+        """
+        return self.is_done_.is_set()
+
+    def wait(self, timeout: Optional[float] = None) -> bool:
+        """
+        Wait for the future to be done.
+
+        Args:
+            timeout (Optional[float]): Maximum time to wait in seconds.
+                If None, wait indefinitely.
+
+        Returns:
+            bool: True if the future is done, False if the timeout was reached.
+        """
+        return self.is_done_.wait(timeout)
+
+    def result(self, timeout: Optional[float] = None) -> T:
+        """
+        Get the result of the future.
+
+        Args:
+            timeout (Optional[float]): Maximum time to wait in seconds.
+                If None, wait indefinitely.
+
+        Returns:
+            T: The result of the future.
+
+        Raises:
+            TimeoutError: If the future is not done within the timeout.
+        """
+        flag = self.wait(timeout)
+        if not flag:
+            raise TimeoutError("Future result not available within timeout")
+        return self.result_
+
+    def set_result(self, result: T) -> None:
+        self.result_ = result
+        self.is_done_.set()
