+++ lmcache/v1/multiprocess/mq.py
@@ -18,6 +18,9 @@ from lmcache.v1.multiprocess.custom_types import (
     get_customized_decoder,
     get_customized_encoder,
 )
+from lmcache.v1.multiprocess.futures import (
+    MessagingFuture,
+)
 from lmcache.v1.multiprocess.protocol import (
     HandlerType,
     RequestType,
@@ -84,57 +87,6 @@ def msgspec_decode(b_obj: bytes, cls: Any) -> Any:
 
 
 # Main classes
-class MessagingFuture(Generic[T]):
-    def __init__(self):
-        self.is_done_ = threading.Event()
-        self.result_ = None
-
-    def query(self) -> bool:
-        """
-        Check if the future is done.
-
-        Returns:
-            bool: True if the future is done, False otherwise.
-        """
-        return self.is_done_.is_set()
-
-    def wait(self, timeout: Optional[float] = None) -> bool:
-        """
-        Wait for the future to be done.
-
-        Args:
-            timeout (Optional[float]): Maximum time to wait in seconds.
-                If None, wait indefinitely.
-
-        Returns:
-            bool: True if the future is done, False if the timeout was reached.
-        """
-        return self.is_done_.wait(timeout)
-
-    def result(self, timeout: Optional[float] = None) -> T:
-        """
-        Get the result of the future.
-
-        Args:
-            timeout (Optional[float]): Maximum time to wait in seconds.
-                If None, wait indefinitely.
-
-        Returns:
-            T: The result of the future.
-
-        Raises:
-            TimeoutError: If the future is not done within the timeout.
-        """
-        flag = self.wait(timeout)
-        if not flag:
-            raise TimeoutError("Future result not available within timeout")
-        return self.result_
-
-    def set_result(self, result: T) -> None:
-        self.result_ = result
-        self.is_done_.set()
-
-
 class MessageQueueClient:
     @dataclass
     class WrappedRequest:
